rename Conifg to config, move generic response and jwtAuthenticationFilter to utils
