+++ skill-tree/src/main/java/com/RDS/skilltree/Endorsement/EndorsementController.java
@@ -1,10 +1,10 @@
 package com.RDS.skilltree.Endorsement;
 
-import com.RDS.skilltree.Common.Response.GenericResponse;
 import com.RDS.skilltree.User.UserModel;
 import com.RDS.skilltree.User.UserRepository;
 import com.RDS.skilltree.models.Skill;
 import com.RDS.skilltree.repositories.SkillRepository;
+import com.RDS.skilltree.utils.GenericResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
