+++ skill-tree/src/main/java/com/RDS/skilltree/config/SecurityConfig.java
@@ -1,11 +1,9 @@
-package com.RDS.skilltree.Config;
+package com.RDS.skilltree.config;
 
 import com.RDS.skilltree.Authentication.AuthEntryPoint;
 import com.RDS.skilltree.Authentication.CustomAccessDeniedHandler;
-import com.RDS.skilltree.Filters.JWTAuthenticationFilter;
 import com.RDS.skilltree.User.UserRoleEnum;
-import java.util.Arrays;
-import java.util.List;
+import com.RDS.skilltree.utils.JWTAuthenticationFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpHeaders;
@@ -19,6 +17,9 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import java.util.Arrays;
+import java.util.List;
 
 @EnableWebSecurity
 @Configuration
