+++ skill-tree/src/main/java/com/RDS/skilltree/utils/GenericResponse.java
@@ -1,4 +1,4 @@
-package com.RDS.skilltree.Common.Response;
+package com.RDS.skilltree.utils;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
