+++ skill-tree/src/main/java/com/RDS/skilltree/utils/JWTAuthenticationFilter.java
@@ -1,14 +1,12 @@
-package com.RDS.skilltree.Filters;
+package com.RDS.skilltree.utils;
 
 import com.RDS.skilltree.Authentication.UserAuthenticationToken;
-import com.RDS.skilltree.utils.JWTUtils;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -19,13 +17,16 @@ import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 import org.springframework.web.util.WebUtils;
 
+import java.io.IOException;
+
 @Slf4j
 public class JWTAuthenticationFilter extends OncePerRequestFilter {
 
     @Value("${cookieName}")
     private String cookieName;
 
-    @Autowired private JWTUtils jwtUtils;
+    @Autowired
+    private JWTUtils jwtUtils;
     private static final String BEARER_PREFIX = "Bearer ";
 
     @Override
